+++ src/components/MetadataCollections/CollectionFilters.js
@@ -1,5 +1,5 @@
 import PropTypes from 'prop-types';
-import { useState } from 'react';
+import { useMemo } from 'react';
 import { FormattedMessage } from 'react-intl';
 
 import {
@@ -10,7 +10,7 @@ import {
 } from '@folio/stripes/components';
 import { CheckboxFilter } from '@folio/stripes/smart-components';
 
-import { useUpdatedFilters } from '../../hooks';
+import { buildFilterState } from '../../util/filterUtils';
 import filterConfig from './filterConfigData';
 
 const CollectionFilters = ({
@@ -24,20 +24,11 @@ const CollectionFilters = ({
   filterData,
   ...props
 }) => {
-  const [filterState, setFilterState] = useState({
-    selected: [],
-    freeContent: [],
-    permitted: [],
-    mdSource: [],
-  });
-
-  useUpdatedFilters({
-    dynamicKey: 'mdSource',
-    filterConfig,
-    filterData,
-    filterState,
-    setFilterState,
-  });
+  const filterState = useMemo(
+    // skip for mdSource filter as it is dynamic and handled separately
+    () => buildFilterState(filterConfig.filter(f => f.name !== 'mdSource')),
+    []
+  );
 
   const renderCheckboxFilter = (key) => {
     const groupFilters = activeFilters[key] || [];
@@ -63,8 +54,8 @@ const CollectionFilters = ({
   };
 
   const renderMetadataSourceFilter = () => {
-    const mdSources = filterData.mdSources;
-    const dataOptions = mdSources.map(mdSource => ({
+    // use dynamic filter values from okapi
+    const dataOptions = (filterData.mdSources || []).map(mdSource => ({
       value: mdSource.id,
       label: mdSource.label,
     }));