+++ src/components/MetadataSources/SourceFilters.js
@@ -1,5 +1,4 @@
 import PropTypes from 'prop-types';
-import { useState } from 'react';
 import { FormattedMessage } from 'react-intl';
 
 import {
@@ -9,8 +8,10 @@ import {
 } from '@folio/stripes/components';
 import { CheckboxFilter } from '@folio/stripes/smart-components';
 
-import { useUpdatedFilters } from '../../hooks';
+import { buildFilterState } from '../../util/filterUtils';
 import filterConfig from './filterConfigData';
+
+const filterState = buildFilterState(filterConfig);
 
 const SourceFilters = ({
   activeFilters = {
@@ -20,19 +21,6 @@ const SourceFilters = ({
   filterHandlers,
   ...props
 }) => {
-  const [filterState, setFilterState] = useState({
-    status: [],
-    selected: [],
-  });
-
-  useUpdatedFilters({
-    dynamicKey: '',
-    filterConfig,
-    filterData: [],
-    filterState,
-    setFilterState,
-  });
-
   const renderCheckboxFilter = (key) => {
     const groupFilters = activeFilters[key] || [];