+++ src/util/filterUtils.js
@@ -1,0 +1,6 @@
+export const buildFilterState = (configs) => Object.fromEntries(
+  configs.map(filter => [
+    filter.name,
+    filter.values.map(({ cql, name }) => ({ value: cql, label: name })),
+  ])
+);